fix: シグナルハンドラはmain関数で登録するように
